Added a README template
Jerry and Leaf,
Your open source project needs a README for others to understand what it is and how to use it. I've helped you by creating a template. You should merge this pull request, then edit this file (edit all the [BLANK SECTIONS]), then commit it so it's saved. It will then appear on your repo main page.